pyproject.toml
[project]
name = "github-todoist-sync"
version = "0.1.0"
description = "Copies GitHub issues to Todoist tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "pygithub>=2.5.0",
    "python-dotenv>=1.0.1",
    "todoist-api-python>=2.1.7",
]

[tool.uv]
dev-dependencies = [
    "bandit>=1.7.9",
    "mypy>=1.11.2",
    "pytest-mock>=3.14.0",
    "pytest>=8.3.2",
    "ruff>=0.6.2",
]

# Ruff has a lot of additional linters that can be really helpful. Only 'E' and 'F' 
# linters are on by default.
[tool.ruff]
exclude = [".git", ".vscode", ".pytest_cache", ".mypy_cache", ".env"]
line-length = 88

[tool.ruff.lint]
ignore = ["B008", "E501", "RET504", "B905", "N806", "N803", "ISC001"]
select = [
    "B",   # https://pypi.org/project/flake8-bugbear/
    "E",   # https://pycodestyle.pycqa.org/en/latest/intro.html#error-codes
    "F",   # https://flake8.pycqa.org/en/latest/user/error-codes.html
    "W",   # https://pycodestyle.pycqa.org/en/latest/intro.html#error-codes
    "I",   # https://pycqa.github.io/isort/
    "N",   # https://github.com/PyCQA/pep8-naming
    "C4",  # https://github.com/adamchainz/flake8-comprehensions
    "EXE", # https://pypi.org/project/flake8-executable/
    "ISC", # https://github.com/flake8-implicit-str-concat/flake8-implicit-str-concat
    "ICN", # https://pypi.org/project/flake8-import-conventions/
    "PIE", # https://pypi.org/project/flake8-pie/
    "PT",  # https://github.com/m-burst/flake8-pytest-style
    "RET", # https://pypi.org/project/flake8-return/
    "SIM", # https://pypi.org/project/flake8-simplify/
    "ERA", # https://pypi.org/project/flake8-eradicate/
    "PLC", # https://beta.ruff.rs/docs/rules/#convention-plc
    "RUF", # https://beta.ruff.rs/docs/rules/#ruff-specific-rules-ruf
    "ARG", # https://beta.ruff.rs/docs/rules/#flake8-unused-arguments-arg
]


# These are bandit settings you might change per project
[tool.bandit]
targets = ["src"]
exclude_dirs = ["tests", ".venv"]

[tool.mypy]
disallow_untyped_defs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
